django-admin/requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
django-admin-interface==0.28.8
django-admin-tools==0.9.3
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
h11==0.14.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pika==1.3.2
pillow==10.4.0
python-slugify==8.0.4
PyYAML==6.0.1
redis==5.0.7
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.37.2
text-unidecode==1.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
